Begin moving files to common file to easily id bugs
