+++ python/plot3.py
@@ -9,6 +9,8 @@ from iree.compiler import compile_str
 from substitute import FuncSubstitute, get_approx_kernel
 import os
 
+from common import test_load
+
 # Configuration for our MNIST model
 NUM_CLASSES = 10
 NUM_ROWS, NUM_COLS = 28, 28
@@ -269,35 +271,6 @@ def test():
     
     return exact_module, func_sub
 
-parent_dir_path = os.path.dirname(os.path.abspath(__file__))
-
-def test_load(load_mlir_path = os.path.join(parent_dir_path, "bin", "output.mlir")):
-    (x_train, y_train, y_train_onehot), (x_test, y_test, y_test_onehot) = load_data()
-    
-    exact_module_path = "mnist_exact_model"
-    # load it back
-    exact_module = tf.saved_model.load(exact_module_path)
-    print("Exact model loaded from saved file.")
-    
-    # Test constructing an approximate kernel
-    print("Creating approximation kernel...")
-    test_kernel = get_approx_kernel(FEATURES_SHAPE, OUTPUT_SHAPE, BATCH_SIZE)
-    print("Successfully created approximation kernel")
-    
-    # Create the function substitution handler
-    func_sub = FuncSubstitute(
-        exact_module=exact_module,
-        approx_kernel=FuncSubstitute.load_mlir_from_file("output.mlir"),
-        input_shape=FEATURES_SHAPE,
-        batch_size=BATCH_SIZE
-    )
-    
-    func_sub.test_comparison = test_comparison.__get__(func_sub)
-    
-    func_sub.test_comparison(
-        x_test, y_test, num_samples=10
-    )
-    
     
 
 if __name__ == "__main__":
